chatproject/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==3.2.16
django-cors-headers==4.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pymongo==3.12.3
pyOpenSSL==24.2.1
pytz==2024.2
redis==5.1.1
service-identity==24.1.0
sqlparse==0.2.4
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.1.0
